Rename resumablemixin file to match class name
