+++ providers/apache/spark/src/airflow/providers/apache/spark/operators/spark_submit.py
@@ -31,7 +31,7 @@ from airflow.providers.common.compat.openlineage.utils.spark import (
 from airflow.providers.common.compat.sdk import BaseOperator, conf
 
 try:
-    from airflow.sdk.bases.resumablemixin import ResumableJobMixin
+    from airflow.sdk import ResumableJobMixin
 except ImportError:
     # Airflow 2 compat.
     # ResumableJobMixin does not exist in Airflow 2, so we need to add a stub to make it
