+++ task-sdk/src/airflow/sdk/__init__.py
@@ -126,7 +126,7 @@ if TYPE_CHECKING:
         cross_downstream,
     )
     from airflow.sdk.bases.operatorlink import BaseOperatorLink
-    from airflow.sdk.bases.resumablemixin import ResumableJobMixin
+    from airflow.sdk.bases.resumablejobmixin import ResumableJobMixin
     from airflow.sdk.bases.sensor import BaseSensorOperator, PokeReturnValue
     from airflow.sdk.bases.skipmixin import SkipMixin
     from airflow.sdk.bases.xcom import BaseXCom
@@ -256,7 +256,7 @@ __lazy_imports: dict[str, str] = {
     "PokeReturnValue": ".bases.sensor",
     "ProductMapper": ".definitions.partition_mappers.product",
     "QuarterWindow": ".definitions.partition_mappers.window",
-    "ResumableJobMixin": ".bases.resumablemixin",
+    "ResumableJobMixin": ".bases.resumablejobmixin",
     "RetryAction": ".definitions.retry_policy",
     "RetryDecision": ".definitions.retry_policy",
     "RetryPolicy": ".definitions.retry_policy",
